Complete the framework: inference, RL, finetuning, eval, config, CLI, docs
Fills in every remaining subsystem so the lifecycle closes end to end.

  models      DiT and AVDiT with 3D rotary embeddings driven by physical
              coordinates, adaLN-single modulation, GQA, and declared
              tensor-parallel plans. Presets from tiny to 14B, all meta-device
              constructible so a large model never has to fit unsharded.
  train       Rectified flow with resolution-shifted timesteps, multi-task
              conditioning, DTensor-safe sharded EMA, WSD schedules, and a
              functional train_step the Trainer orchestrates rather than owns.
  data        Deterministic synthetic source, mmap shard format with atomic
              commit and sha256, resolution/duration bucketing, sequence
              packing, and a loader that shards on data_rank alone.
  checkpoint  DCP with async save, resharding load, retention, and separate
              safetensors export for release artifacts.
  infer       Five samplers verified to integrate the rectified-flow field to
              within 5e-7, CFG with rescale and adaptive projected guidance,
              and conditioning builders that produce byte-identical layouts to
              the training path.
  finetune    LoRA and DoRA implemented natively on DTensor, control and image
              adapters, all exactly identity at initialisation.
  rl          Flow-GRPO with an ODE-to-SDE conversion that returns exact
              transition log-probs, group-relative advantage, MixGRPO windows,
              and a DPO objective that drops into the supervised trainer.
  eval        Dependency-free metrics plus gated learned backends that refuse
              to substitute silently, and reports that refuse to compare
              mismatched pins.
  codecs      Invertible reference codecs for CI, tiled adapters for real VAEs.
  config      Dataclasses over YAML with _base_ composition and dotted
              overrides. Unknown keys are fatal, because a silently ignored
              typo is a wasted cluster run.
  cli         train, finetune, rl, generate, eval, simulate, plan, checkpoint,
              data, info.

Ships the open-source scaffolding: contributor guide, code of conduct, security
and governance policy, issue forms, dependabot, and five workflows. The simulate
workflow is the notable one — it prices committed reference plans at world sizes
8, 64, 512 and 1024 on a CPU runner and fails the build when a plan stops
fitting or its predicted utilisation regresses.

Cross-subsystem fixes from integration: export gather_tokens, add the rewards
extra the RL path tells users to install, reconcile the checkpoint retention
keyword with the config schema, relax mypy's subclassing rule (torch is followed
as Any, which would otherwise reject every model class), and correct two doc
snippets that referenced APIs which do not exist.

No dataset names, sources, quotas or tuned thresholds anywhere: the data layer
ships documented templates whose quality gates default to disabled.

110 tests pass, ruff is clean across source and tests, all 19 shipped configs
load and round-trip, and importing avgen pulls in no optional dependency.

+++ src/avgen/train/schedule.py
@@ -305,7 +305,7 @@ def _warmup_stable_decay(
         raise ValueError(
             f"decay_shape must be one of 1-sqrt, linear, cosine; got {shape!r}"
         )
-    decay_steps = max(int(round(total_steps * fraction)), 1)
+    decay_steps = max(round(total_steps * fraction), 1)
     decay_start = max(total_steps - decay_steps, warmup_steps)
     span = max(total_steps - decay_start, 1)
